requirements.txt
google-generativeai
pdfplumber
requests
python-dotenv